Siga and MainFrame inline docs done - Fixed AI functionality bugs on Siga - Enhanced the MainFrame capabilities with options like quit and return
